+++ CHANGELOG.md
@@ -3,13 +3,40 @@
 
 
 
-## [1.1.1] - 2022-02-25
-[Full Changelog](https://github.com/TechNobre/PowerUtils.Net.Primitives/compare/v1.1.0...v1.1.1)
+## [1.2.0] - 2022-03-02
+[Full Changelog](https://github.com/TechNobre/PowerUtils.Net.Primitives/compare/v1.1.1...v1.2.0)
 
 
 ### New Features
 
-- Fixed link for status code UNPROCESSABLE_ENTITY  422;
+- Added constant for Status Code I'm a teapot 418;
+- Added constant for Status Code Too Early 425;
+- Added constant for Status Code Precondition Required 428;
+- Added constant for Status Code Too Many Requests 429;
+- Added constant for Status Code Request Header Fields Too Large 431;
+- Added constant for Status Code Unavailable For Legal Reasons 451;
+- Added constant for Status Code Variant Also Negotiates 506;
+- Added constant for Status Code Insufficient Storage 507;
+- Added constant for Status Code Loop Detected 508;
+- Added constant for Status Code Not Extended 510;
+- Added constant for Status Code Network Authentication Required 511;
+- Added utils `int.GetStatusCodeLinkOrDefault()`;
+
+
+### Fixed
+
+- Fixed methods documentation;
+
+
+
+
+## [1.1.1] - 2022-02-25
+[Full Changelog](https://github.com/TechNobre/PowerUtils.Net.Primitives/compare/v1.1.0...v1.1.1)
+
+
+### Fixed
+
+- Fixed link for status code Unprocessable Entity 422;
 
 
 
@@ -20,7 +47,7 @@
 
 ### New Features
 
-- Added constant for Status Code UNPROCESSABLE ENTITY 422;
+- Added constant for Status Code Unprocessable Entity 422;
 
 
 ### Enhancements